Added MQTT test, fixed wrong POST body, renamed various files and bundle
