+++ kura/org.eclipse.kura.rest.service.listing.provider/src/main/java/org/eclipse/kura/internal/rest/service/listing/provider/RestServiceListingProvider.java
@@ -78,7 +78,6 @@ public class RestServiceListingProvider {
             logger.debug(DEBUG_MESSSAGE, "serviceListing/v1/sortedList");
 
             BundleContext context = FrameworkUtil.getBundle(RestServiceListingProvider.class).getBundleContext();
-            
             List<String> resultDTO = getAllServices(context);
 
             return new SortedServiceListDTO(resultDTO);
