+++ kura/test/org.eclipse.kura.rest.service.listing.provider.test/META-INF/MANIFEST.MF
@@ -8,7 +8,6 @@ Bundle-License: Eclipse Public License v2.0
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 Bundle-ActivationPolicy: lazy
 Import-Package: javax.annotation;version="1.2.0",
- org.eclipse.kura.configuration;version="1.2.0",
  org.eclipse.kura.core.testutil.requesthandler;version="1.1.0",
  org.eclipse.kura.util.wire.test;version="1.1.0",
  org.junit;version="[4.12.0,5.0.0)",
